backend/requirements.txt
# Core FastAPI and web server requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Authentication and security requirements
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-jwt>=3.3.0
PyJWT>=2.8.0

# Database requirements (SQLite for local, PostgreSQL for Docker)
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.0; platform_system != "Windows"
psycopg2>=2.9.0; platform_system == "Windows"

# Image processing and OCR (Windows compatible versions)
Pillow>=9.0.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0

# Date/time utilities
python-dateutil>=2.8.0

# Scheduler for automatic overdue notifications
apscheduler>=3.11.0
pytz>=2023.3

# Logging and HTTP client
httpx>=0.25.0

# Additional Windows compatibility
wheel
setuptools
